refactor: Move query plan normalizer to arrow_util
