+++ arrow_util/Cargo.toml
@@ -15,7 +15,10 @@ chrono = { version = "0.4", default-features = false }
 comfy-table = { version = "6.1", default-features = false }
 hashbrown = { workspace = true }
 num-traits = "0.2"
+once_cell = { version = "1.17", features = ["parking_lot"] }
+regex = "1.7.0"
 snafu = "0.7"
+uuid = "1"
 workspace-hack = { version = "0.1", path = "../workspace-hack" }
 
 [dev-dependencies]
