+++ querier/Cargo.toml
@@ -57,4 +57,3 @@ iox_tests = { path = "../iox_tests" }
 mutable_batch_lp = { path = "../mutable_batch_lp" }
 object_store_metrics = { path = "../object_store_metrics" }
 test_helpers = { path = "../test_helpers" }
-test_helpers_end_to_end = { path = "../test_helpers_end_to_end" }
